+++ components/angular/projects/ng-particles/src/lib/ng-particles.component.ts
@@ -1,10 +1,8 @@
 import { AfterViewInit, Component, EventEmitter, Inject, Input, OnDestroy, Output, PLATFORM_ID } from '@angular/core';
 import { isPlatformServer } from '@angular/common';
 import { EMPTY, from, mergeMap, Subject, takeUntil } from 'rxjs';
-
 import { tsParticles } from 'tsparticles-engine';
 import type { Container, Engine } from 'tsparticles-engine';
-
 import { IParticlesProps } from './ng-particles.module';
 
 @Component({
@@ -14,13 +12,16 @@ import { IParticlesProps } from './ng-particles.module';
 export class NgParticlesComponent implements AfterViewInit, OnDestroy {
     @Input() options?: IParticlesProps;
     @Input() url?: string;
-    @Input() id?: string;
+    @Input() id: string;
     @Input() particlesInit?: (engine: Engine) => Promise<void>;
     @Output() particlesLoaded: EventEmitter<Container> = new EventEmitter<Container>();
 
     private destroy$ = new Subject<void>();
+    private container?: Container;
 
-    constructor(@Inject(PLATFORM_ID) private platformId: string) {}
+    constructor(@Inject(PLATFORM_ID) protected platformId: string) {
+        this.id = 'tsparticles';
+    }
 
     public ngAfterViewInit(): void {
         if (isPlatformServer(this.platformId)) {
@@ -28,6 +29,7 @@ export class NgParticlesComponent implements AfterViewInit, OnDestroy {
         }
 
         const cb = (container?: Container) => {
+            this.container = container;
             this.particlesLoaded.emit(container);
         };
 
@@ -35,17 +37,9 @@ export class NgParticlesComponent implements AfterViewInit, OnDestroy {
             .pipe(
                 mergeMap(() => {
                     if (this.url) {
-                        if (this.id) {
-                            return tsParticles.loadJSON(this.id, this.url);
-                        } else {
-                            return tsParticles.loadJSON(this.url);
-                        }
+                        return tsParticles.loadJSON(this.id, this.url);
                     } else if (this.options) {
-                        if (this.id) {
-                            return tsParticles.load(this.id, this.options);
-                        } else {
-                            return tsParticles.load(this.options);
-                        }
+                        return tsParticles.load(this.id, this.options);
                     } else {
                         console.error('You must specify options or url to load tsParticles');
                         return EMPTY;
@@ -57,6 +51,7 @@ export class NgParticlesComponent implements AfterViewInit, OnDestroy {
     }
 
     public ngOnDestroy(): void {
+        this.container?.destroy();
         this.destroy$.next();
     }
 }
@@ -68,7 +63,13 @@ export class NgParticlesComponent implements AfterViewInit, OnDestroy {
 export class ParticlesComponent extends NgParticlesComponent {
     @Input() override options?: IParticlesProps;
     @Input() override url?: string;
-    @Input() override id?: string;
+    @Input() override id: string;
     @Input() override particlesInit?: (engine: Engine) => Promise<void>;
     @Output() override particlesLoaded: EventEmitter<Container> = new EventEmitter<Container>();
+
+    constructor(@Inject(PLATFORM_ID) protected override platformId: string) {
+        super(platformId);
+
+        this.id = 'tsparticles';
+    }
 }