+++ demo/jquery/app.js
@@ -22,6 +22,7 @@ app.use(express.static('./public'));
 app.use("/fontawesome", express.static("./node_modules/@fortawesome/fontawesome-free"));
 app.use("/jsoneditor", express.static("./node_modules/jsoneditor/dist"));
 app.use("/tsparticles", express.static("./node_modules/tsparticles"));
+app.use("/demo-configs", express.static("./node_modules/tsparticles-demo-configs"));
 app.use("/jquery-particles", express.static("./node_modules/jquery-particles/dist"));
 app.use("/preset-links", express.static("./node_modules/tsparticles-preset-links"));
 app.use("/stats.ts", express.static("./node_modules/stats.ts/"));