+++ engine/src/Options/Classes/Interactivity/Events/Events.ts
@@ -4,6 +4,7 @@ import { HoverEvent } from "./HoverEvent";
 import type { IEvents } from "../../../Interfaces/Interactivity/Events/IEvents";
 import type { IOptionLoader } from "../../../Interfaces/IOptionLoader";
 import type { RecursivePartial } from "../../../../Types/RecursivePartial";
+import { ResizeEvent } from "./ResizeEvent";
 import type { SingleOrMultiple } from "../../../../Types/SingleOrMultiple";
 import { executeOnSingleOrMultiple } from "../../../../Utils/Utils";
 
@@ -21,7 +22,7 @@ export class Events implements IEvents, IOptionLoader<IEvents> {
         this.onClick = new ClickEvent();
         this.onDiv = new DivEvent();
         this.onHover = new HoverEvent();
-        this.resize = true;
+        this.resize = new ResizeEvent();
     }
 
     /**
@@ -96,8 +97,10 @@ export class Events implements IEvents, IOptionLoader<IEvents> {
 
         this.onHover.load(data.onHover ?? data.onhover);
 
-        if (data.resize !== undefined) {
-            this.resize = data.resize;
+        if (typeof data.resize === "boolean") {
+            this.resize.enable = data.resize;
+        } else {
+            this.resize.load(data.resize);
         }
     }
 }