+++ interactions/particles/links/src/Linker.ts
@@ -66,6 +66,7 @@ export class Linker extends ParticlesInteractorBase {
     }
 
     init(): void {
+        this.linkContainer.particles.linksColor = undefined;
         this.linkContainer.particles.linksColors = new Map<string, IRgb | string | undefined>();
     }