+++ components/vue-alt/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.6.0](https://github.com/matteobruni/tsparticles/compare/particles.vue@3.5.3...particles.vue@3.6.0) (2022-12-06)
+
+**Note:** Version bump only for package particles.vue
+
 ## [3.5.3](https://github.com/matteobruni/tsparticles/compare/particles.vue@3.5.2...particles.vue@3.5.3) (2022-11-07)
 
 **Note:** Version bump only for package particles.vue