+++ demo/vue3/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.6.0](https://github.com/matteobruni/tsparticles/compare/vue3-demo@0.5.3...vue3-demo@0.6.0) (2022-12-06)
+
+**Note:** Version bump only for package vue3-demo
+
 ## [0.5.3](https://github.com/matteobruni/tsparticles/compare/vue3-demo@0.5.2...vue3-demo@0.5.3) (2022-11-07)
 
 **Note:** Version bump only for package vue3-demo