+++ websites/confetti.js.org/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.5.0](https://github.com/matteobruni/confetti/compare/@tsparticles/confetti-website@1.4.3...@tsparticles/confetti-website@1.5.0) (2022-12-06)
+
+**Note:** Version bump only for package @tsparticles/confetti-website
+
 ## [1.4.3](https://github.com/matteobruni/confetti/compare/@tsparticles/confetti-website@1.4.2...@tsparticles/confetti-website@1.4.3) (2022-11-07)
 
 **Note:** Version bump only for package @tsparticles/confetti-website